LPS-158489 Inline InternetAddressUtil toString method into mail-reader-service since it is the only usage
