+++ modules/apps/archived/mail-reader-service/src/main/java/com/liferay/mail/reader/internal/imap/IMAPAccessor.java
@@ -20,12 +20,12 @@ import com.liferay.mail.reader.constants.MailConstants;
 import com.liferay.mail.reader.exception.MailException;
 import com.liferay.mail.reader.exception.NoSuchMessageException;
 import com.liferay.mail.reader.internal.util.HtmlContentUtil;
+import com.liferay.mail.reader.internal.util.InternetAddressUtil;
 import com.liferay.mail.reader.model.Account;
 import com.liferay.mail.reader.model.MailFile;
 import com.liferay.mail.reader.service.AttachmentLocalServiceUtil;
 import com.liferay.mail.reader.service.FolderLocalServiceUtil;
 import com.liferay.mail.reader.service.MessageLocalServiceUtil;
-import com.liferay.petra.mail.InternetAddressUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
