+++ modules/apps/archived/mail-reader-service/src/main/java/com/liferay/mail/reader/internal/mailbox/IMAPMailbox.java
@@ -23,6 +23,7 @@ import com.liferay.mail.reader.internal.attachment.DefaultAttachmentHandler;
 import com.liferay.mail.reader.internal.imap.IMAPAccessor;
 import com.liferay.mail.reader.internal.imap.IMAPConnection;
 import com.liferay.mail.reader.internal.util.AccountLock;
+import com.liferay.mail.reader.internal.util.InternetAddressUtil;
 import com.liferay.mail.reader.model.Account;
 import com.liferay.mail.reader.model.Attachment;
 import com.liferay.mail.reader.model.Folder;
@@ -33,7 +34,6 @@ import com.liferay.mail.reader.service.AccountLocalServiceUtil;
 import com.liferay.mail.reader.service.AttachmentLocalServiceUtil;
 import com.liferay.mail.reader.service.FolderLocalServiceUtil;
 import com.liferay.mail.reader.service.MessageLocalServiceUtil;
-import com.liferay.petra.mail.InternetAddressUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
